requirements.txt
gradio==4.12.0
markdown==3.5
Pillow==10.0.0
numpy==1.24.3
emoji==2.8.0 